Rename training to autodoc
